requirements.txt
pygame>=2.5.0
PyOpenGL>=3.1.0
PyOpenGL-accelerate>=3.1.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0

